Downloads handled by bridge instead of mod
The bridge will be what handles downloading the mods instead of relying
on the mod-side to write the files itself. The bridge communicates the
progress of the file, however.

Also introduces a basic state sharing framework but at the moment is
used to share where mods should be downloaded to and a to-be-implemented
switch to enable unix compatibility should the bridge not need to fix
the given path download path. State is initially given by the mod on
connect.

Additionally a couple of miscellaneous things were changed that didn't
age too well or were changed to be reused.